feat: add seed script to generate initial auctions data
- Added a seed script (`seed.ts`) to generate initial data for auctions.
- The script uses back-end types to ensure consistency with the server.
- Updated the `package.json` to include the seed script in the `dev` command.
- Updated the `tsconfig.json` to include the `@/seeds/*` import path.
- Added a new checkbox column to the auction lots table for row selection.

Closes #123
