+++ src/features/pre-auction/auction-lots/components/columns.tsx
@@ -8,10 +8,37 @@ import { AuctionEntity } from '../types/auction.entity.ts'
 import { DataTableColumnHeader } from './data-table-column-header'
 import { DataTableRowActions } from './data-table-row-actions'
 
+import { Checkbox } from '@/src/components/ui/checkbox.jsx'
 import Link from 'next/link'
 import React from 'react'
 
 export const columns: ColumnDef<AuctionEntity>[] = [
+   {
+      id: 'select',
+      header: ({ table }) => (
+         <Checkbox
+            checked={
+               table.getIsAllPageRowsSelected() ||
+               (table.getIsSomePageRowsSelected() && 'indeterminate')
+            }
+            onCheckedChange={(value) =>
+               table.toggleAllPageRowsSelected(!!value)
+            }
+            aria-label="Select all"
+            className="translate-y-[2px]"
+         />
+      ),
+      cell: ({ row }) => (
+         <Checkbox
+            checked={row.getIsSelected()}
+            onCheckedChange={(value) => row.toggleSelected(!!value)}
+            aria-label="Select row"
+            className="translate-y-[2px]"
+         />
+      ),
+      enableSorting: false,
+      enableHiding: false
+   },
    {
       accessorKey: 'auctionDate',
       header: ({ column }) => (
