+++ tsconfig.json
@@ -27,7 +27,8 @@
          "@/styles/*": ["./src/styles/*"],
          "@/hooks/*": ["./src/hooks/*"],
          "@/utils/*": ["./src/utils/*"],
-         "@/svgs/*": ["./src/components/svgs/*"]
+         "@/svgs/*": ["./src/components/svgs/*"],
+         "@/seeds/*": ["./src/seeds/*"]
       }
    },
    "include": [
